feat(v3-10): the ALTER refusals carry an ALTER phrase and Spark's classes
The v3 opt-in refusal is composed for the ALTER door instead of borrowing the CREATE
door's, whose tail ("default create stays format v2") is a create-only clause; the conf
key stays one shared const, which is the half every door's pin asserts.

The request is parsed as a SIGNED integer and no longer trimmed, so `-1` and `0` refuse as
downgrades naming both versions (Spark: "Cannot downgrade v2 table to v-1") and `3.0`,
`''` and `' 3 '` refuse as unparsable (Spark: For input string). `+3` upgrades on both.
The ANSI map row moves under the create.rs bullet where the tests live; STATUS is
compacted back under its ceiling.

+++ crates/repark-iceberg/src/write/format_version.rs
@@ -41,12 +41,14 @@ pub async fn set_properties_and_format_version<S: BuildHasher>(
 }
 
 #[allow(clippy::missing_errors_doc)]
-pub async fn current_format_version(catalog: &dyn Catalog, ident: &TableIdent) -> Result<u8> {
-    Ok(catalog.load_table(ident).await?.metadata().format_version() as u8)
+pub async fn current_format_version(catalog: &dyn Catalog, ident: &TableIdent) -> Result<i64> {
+    Ok(i64::from(
+        catalog.load_table(ident).await?.metadata().format_version() as u8,
+    ))
 }
 
 #[allow(clippy::missing_errors_doc)]
-pub fn format_version_from_number(number: u8) -> Result<FormatVersion> {
+pub fn format_version_from_number(number: i64) -> Result<FormatVersion> {
     match number {
         1 => Ok(FormatVersion::V1),
         2 => Ok(FormatVersion::V2),
